refactor: 중복 애니메이션을 shared로 이동하고 인라인 스타일 제거
