+++ apps/web/src/shared/ui/animations/index.ts
@@ -1,0 +1,1 @@
+export { blinkingText } from './loading.css';
