+++ apps/web/src/widgets/home/ui/ExpenseSummary/ExpenseSummary.css.ts
@@ -1,4 +1,4 @@
-import { style, keyframes } from '@vanilla-extract/css';
+import { style } from '@vanilla-extract/css';
 import { vars } from '@/shared/ui/theme.css';
 
 export const container = style({
@@ -20,18 +20,10 @@ export const summaryRow = style({
 
 export const countText = style({
   flexShrink: 0,
+  color: vars.color.text.secondary,
 });
 
 export const amountText = style({
   flexShrink: 0,
-});
-
-const blink = keyframes({
-  '0%': { opacity: 1 },
-  '50%': { opacity: 0.5 },
-  '100%': { opacity: 1 },
-});
-
-export const blinkingText = style({
-  animation: `${blink} 1.5s ease-in-out infinite`,
+  color: vars.color.text.primary,
 });
