+++ apps/web/src/widgets/home/ui/ExpenseSummary/ExpenseSummary.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Text } from '@/shared/ui/text';
-import { vars } from '@/shared/ui/theme.css';
+import { blinkingText } from '@/shared/ui/animations';
 import * as styles from './ExpenseSummary.css';
 
 export interface ExpenseSummaryProps {
@@ -19,13 +19,12 @@ export const ExpenseSummary = ({
   return (
     <div className={styles.container}>
       <div className={styles.summaryRow}>
-        <Text variant='b2' color={vars.color.text.secondary} className={styles.countText}>
+        <Text variant='b2' className={styles.countText}>
           {count}건의 소비
         </Text>
         <Text
           variant='h1'
-          color={vars.color.text.primary}
-          className={`${styles.amountText} ${isLoading || isFetching ? styles.blinkingText : ''}`}>
+          className={`${styles.amountText} ${isLoading || isFetching ? blinkingText : ''}`}>
           {totalAmount.toLocaleString()}원
         </Text>
       </div>
