+++ apps/web/src/widgets/home/ui/MonthlyExpenseInfo/MonthlyExpenseInfo.css.ts
@@ -1,4 +1,4 @@
-import { style, keyframes } from '@vanilla-extract/css';
+import { style } from '@vanilla-extract/css';
 import { vars } from '@/shared/ui/theme.css';
 
 export const container = style({
@@ -26,14 +26,4 @@ export const amountText = style({
 export const viewToggleWrapper = style({
   display: 'flex',
   marginBottom: '0.3rem',
-});
-
-const blink = keyframes({
-  '0%': { opacity: 1 },
-  '50%': { opacity: 0.5 },
-  '100%': { opacity: 1 },
-});
-
-export const blinkingText = style({
-  animation: `${blink} 1.5s ease-in-out infinite`,
 });
